+++ FM-Radio@hilyxx/files/FM-Radio@hilyxx/radio.js
@@ -1,6 +1,8 @@
 // === IMPORTS & CONSTANTS ===
 imports.gi.versions.Gst = "1.0";
 const Gst = imports.gi.Gst;
+const Gio = imports.gi.Gio;
+const GLib = imports.gi.GLib;
 const St = imports.gi.St;
 const Clutter = imports.gi.Clutter;
 
@@ -44,7 +46,6 @@ function createControlButtons(player, pr) {
     box.add_child(icon);
     box.add_child(next);
 
-    // Expose this icon to the applet so it can be updated
     pr.playStopIcon = icon;
 
     next.connect("button-press-event", () => {
@@ -79,11 +80,63 @@ const RadioPlayer = class RadioPlayer {
     constructor(channel) {
         this.channel = channel;
         this.playing = false;
+        this.isReconnecting = false;
         this.volume = DEFAULT_VOLUME;
         this.playbin = null;
         this.sink = null;
         this.onError = null;
         this.onTagChanged = null;
+        
+        this._retryTimerId = null;
+
+        // System listener to instantly detect Wi-Fi/Network changes
+        this.networkMonitor = Gio.NetworkMonitor.get_default();
+        this.networkMonitorId = this.networkMonitor.connect("network-changed", (monitor, available) => {
+            if (!this.playing) return;
+            
+            if (!available) {
+                this._startReconnectLoop();
+            } else if (this.isReconnecting) {
+                this._forceReconnectTick();
+            }
+        });
+    }
+
+    _clearRetry() {
+        if (this._retryTimerId) {
+            GLib.source_remove(this._retryTimerId);
+            this._retryTimerId = null;
+        }
+        this.isReconnecting = false;
+    }
+
+    // Starts the reconnection loop (every 5 seconds)
+    _startReconnectLoop() {
+        if (!this.playing || this.isReconnecting) return;
+        
+        this.isReconnecting = true;
+        
+        if (this.playbin) {
+            this.playbin.set_state(Gst.State.NULL);
+        }
+        
+        if (this.onError != null) this.onError(); 
+
+        this._retryTimerId = GLib.timeout_add(GLib.PRIORITY_DEFAULT, 5000, () => {
+            this._forceReconnectTick();
+            return GLib.SOURCE_CONTINUE;
+        });
+    }
+
+    // Attempts to restart the stream
+    _forceReconnectTick() {
+        if (this.playing && this.networkMonitor.get_network_available()) {
+            if (this.playbin) {
+                this.playbin.set_state(Gst.State.NULL);
+                this.playbin.set_property("uri", this.channel.getLink());
+                this.playbin.set_state(Gst.State.PLAYING);
+            }
+        }
     }
 
     _initPipeline() {
@@ -106,6 +159,7 @@ const RadioPlayer = class RadioPlayer {
     }
 
     play() {
+        this._clearRetry();
         this._initPipeline();
         this.playbin.set_state(Gst.State.PLAYING);
         this.playing = true;
@@ -126,10 +180,19 @@ const RadioPlayer = class RadioPlayer {
     }
 
     stop() {
+        this._clearRetry();
         if (this.playbin) {
             this.playbin.set_state(Gst.State.NULL);
         }
         this.playing = false;
+    }
+
+    destroy() {
+        this.stop();
+        if (this.networkMonitorId) {
+            this.networkMonitor.disconnect(this.networkMonitorId);
+            this.networkMonitorId = 0;
+        }
     }
 
     next() {
@@ -218,13 +281,15 @@ const RadioPlayer = class RadioPlayer {
             }
 
             case Gst.MessageType.STREAM_START:
+                this._clearRetry();
                 if (this.onTagChanged != null) this.onTagChanged();
                 break;
 
             case Gst.MessageType.EOS:
             case Gst.MessageType.ERROR:
-                this.stop();
-                if (this.onError != null) this.onError();
+                if (this.playing) {
+                    this._startReconnectLoop();
+                }
                 break;
             default:
                 break;